refactor(sdk-core): use @hypercerts-org/lexicon package for lexicon definitions
- Add index.ts to lexicon package exporting HYPERCERT_LEXICONS and HYPERCERT_COLLECTIONS
- Move hypercertCollection.json to lexicons/org/hypercerts/collection.json
- Add @hypercerts-org/lexicon as workspace dependency to sdk-core
- Update sdk-core to import lexicons from @hypercerts-org/lexicon
- Remove duplicate lexicon JSON files from sdk-core
- Update HYPERCERT_COLLECTIONS.RECORD to CLAIM to match new lexicon naming
