+++ packages/sdk-core/package.json
@@ -76,6 +76,7 @@
     "@atproto/lexicon": "^0.5.1",
     "@atproto/oauth-client": "^0.5.9",
     "@atproto/oauth-client-node": "^0.3.10",
+    "@hypercerts-org/lexicon": "workspace:*",
     "eventemitter3": "^5.0.1",
     "zod": "^3.24.4"
   }
