--- packages/sdk-core/src/lexicons/hypercerts/defs.json
@@ -1,24 +1,0 @@
-{
-  "lexicon": 1,
-  "id": "org.hypercerts.defs",
-  "defs": {
-    "uri": {
-      "type": "string",
-      "format": "uri",
-      "maxGraphemes": 1000,
-      "description": "URI to external data"
-    },
-    "smallBlob": {
-      "type": "blob",
-      "accept": ["*/*"],
-      "maxSize": 10485760,
-      "description": "Blob to external data (up to 10MB)"
-    },
-    "largeBlob": {
-      "type": "blob",
-      "accept": ["*/*"],
-      "maxSize": 104857600,
-      "description": "Blob to external data (up to 100MB)"
-    }
-  }
-}
