--- packages/sdk-core/src/lexicons/hypercerts/hypercert.json
@@ -1,103 +1,0 @@
-{
-  "lexicon": 1,
-  "id": "org.hypercerts.claim.record",
-  "defs": {
-    "main": {
-      "type": "record",
-      "description": "A hypercert record tracking impact work.",
-      "key": "any",
-      "record": {
-        "type": "object",
-        "required": ["title", "description", "createdAt", "workScope", "workTimeframeFrom", "workTimeFrameTo"],
-        "properties": {
-          "title": {
-            "type": "string",
-            "description": "Title of the hypercert",
-            "maxLength": 256
-          },
-          "shortDescription": {
-            "type": "string",
-            "description": "(Optional) short blurb of the impact work done.",
-            "maxLength": 3000,
-            "maxGraphemes": 300
-          },
-          "description": {
-            "type": "string",
-            "description": "A description of the impact work done.",
-            "maxLength": 30000,
-            "maxGraphemes": 3000
-          },
-          "image": {
-            "type": "union",
-            "refs": ["org.hypercerts.defs#uri", "org.hypercerts.defs#smallBlob"],
-            "description": "The hypercert visual representation as a URI or blob"
-          },
-          "workScope": {
-            "type": "string",
-            "description": "Scope of the work performed",
-            "maxLength": 5000,
-            "maxGraphemes": 1000
-          },
-          "workTimeframeFrom": {
-            "type": "string",
-            "format": "datetime",
-            "description": "When the work began"
-          },
-          "workTimeFrameTo": {
-            "type": "string",
-            "format": "datetime",
-            "description": "When the work ended"
-          },
-          "evidence": {
-            "type": "array",
-            "description": "Supporting evidence, documentation, or external data URIs",
-            "items": {
-              "type": "object",
-              "required": ["uri"],
-              "properties": {
-                "uri": {
-                  "type": "string",
-                  "format": "uri",
-                  "description": "Evidence resource URI"
-                },
-                "title": {
-                  "type": "string",
-                  "maxLength": 256,
-                  "description": "Optional title to describe the nature of the evidence"
-                },
-                "description": {
-                  "type": "string",
-                  "maxLength": 3000,
-                  "maxGraphemes": 300,
-                  "description": "Optional short description explaining what this evidence demonstrates or proves"
-                }
-              }
-            },
-            "maxLength": 100
-          },
-          "contributions": {
-            "type": "array",
-            "description": "A strong reference to the contributions done to create the impact in the hypercerts. The record referenced must conform with the lexicon org.hypercerts.claim.contributions",
-            "items": {
-              "type": "ref",
-              "ref": "com.atproto.repo.strongRef"
-            }
-          },
-          "rights": {
-            "ref": "com.atproto.repo.strongRef",
-            "description": "A strong reference to the rights that this hypercert has. The record referenced must conform with the lexicon org.hypercerts.claim.rights"
-          },
-          "location": {
-            "ref": "com.atproto.repo.strongRef",
-            "description": "A strong reference to the location where the work for done hypercert was located. The record referenced must conform with the lexicon org.hypercerts.claim.location"
-          },
-          "createdAt": {
-            "type": "string",
-            "format": "datetime",
-            "description": "Client-declared timestamp when this record was originally created"
-          }
-        }
-      }
-    }
-  }
-}
