--- packages/sdk-core/src/lexicons/hypercerts/hypercertContribution.json
@@ -1,56 +1,0 @@
-{
-  "lexicon": 1,
-  "id": "org.hypercerts.claim.contribution",
-  "defs": {
-    "main": {
-      "type": "record",
-      "description": "A contribution made toward a hypercert's impact.",
-      "key": "any",
-      "record": {
-        "type": "object",
-        "required": ["contributors", "role", "createdAt"],
-        "properties": {
-          "hypercert": {
-            "type": "ref",
-            "ref": "com.atproto.repo.strongRef",
-            "description": "A strong reference to the hypercert this contribution is for. The record referenced must conform with the lexicon org.hypercerts.claim.record"
-          },
-          "role": {
-            "type": "string",
-            "description": "Role or title of the contributor(s).",
-            "maxLength": 100
-          },
-          "contributors": {
-            "type": "array",
-            "description": "List of DIDs identifying the contributors. If multiple are stored in the same hypercertContribution, then they would have the exact same role.",
-            "items": {
-              "type": "string",
-              "format": "did"
-            }
-          },
-          "description": {
-            "type": "string",
-            "description": "What the contribution concretely achieved",
-            "maxLength": 2000,
-            "maxGraphemes": 500
-          },
-          "workTimeframeFrom": {
-            "type": "string",
-            "format": "datetime",
-            "description": "When this contribution started. This should be a subset of the hypercert timeframe."
-          },
-          "workTimeframeTo": {
-            "type": "string",
-            "format": "datetime",
-            "description": "When this contribution finished.  This should be a subset of the hypercert timeframe."
-          },
-          "createdAt": {
-            "type": "string",
-            "format": "datetime",
-            "description": "Client-declared timestamp when this record was originally created"
-          }
-        }
-      }
-    }
-  }
-}
