--- packages/sdk-core/src/lexicons/hypercerts/hypercertMeasurement.json
@@ -1,60 +1,0 @@
-{
-  "lexicon": 1,
-  "id": "org.hypercerts.claim.measurement",
-  "defs": {
-    "main": {
-      "type": "record",
-      "description": "External measurement data supporting a hypercert claim",
-      "key": "tid",
-      "record": {
-        "type": "object",
-        "required": ["hypercert", "measurers", "metric", "value", "createdAt"],
-        "properties": {
-          "hypercert": {
-            "type": "ref",
-            "ref": "com.atproto.repo.strongRef",
-            "description": "A strong reference to the hypercert that this measurement is for. The record referenced must conform with the lexicon org.hypercerts.claim.record"
-          },
-          "measurers": {
-            "type": "array",
-            "description": "DIDs of the entity (or entities) that measured this data. ",
-            "items": {
-              "type": "string",
-              "format": "did"
-            },
-            "maxLength": 100
-          },
-          "metric": {
-            "type": "string",
-            "description": "The metric being measured",
-            "maxLength": 500
-          },
-          "value": {
-            "type": "string",
-            "description": "The measured value",
-            "maxLength": 500
-          },
-          "measurementMethodURI": {
-            "type": "string",
-            "format": "uri",
-            "description": "URI to methodology documentation, standard protocol, or measurement procedure"
-          },
-          "evidenceURI": {
-            "type": "array",
-            "description": "URIs to supporting evidence or data",
-            "items": {
-              "type": "string",
-              "format": "uri"
-            },
-            "maxLength": 50
-          },
-          "createdAt": {
-            "type": "string",
-            "format": "datetime",
-            "description": "Client-declared timestamp when this record was originally created"
-          }
-        }
-      }
-    }
-  }
-}
